requirements.txt
shortuuid==1.0.1
validators==0.18.2
#sentence_transformers==2.0.0
nltk==3.6.2
dash_html_components==1.1.4
requests==2.26.0
finvizfinance==0.9.9
pandas==1.3.1
tweepy==3.10.0
newspaper3k==0.2.8
dash==1.21.0
sumy==0.8.1
dash_core_components==1.17.1
gensim==4.1.0
joblib==1.0.1
dash_bootstrap_components==0.13.0
wordcloud==1.8.1
yfinance==0.1.63
scipy==1.7.1
dash_table==4.12.0
snscrape==0.3.4
spacy==3.1.1
psaw==0.1.0
cassandra_driver==3.25.0
plotly==5.3.1
pyLDAvis==3.3.1
gunicorn==20.0.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz
